Compiler warnings and code tidying

+++ gtkoptions.c
@@ -1418,9 +1418,9 @@ static void OptionsOK(GtkWidget *pw, optionswidget *pow)
 	  if (gtk_toggle_button_get_active( GTK_TOGGLE_BUTTON( pow->apwDice[i])))
 		  break;
   }
-  if (i < RNG_FILE && i != rngCurrent ) {
+  if (i < RNG_FILE && i != (unsigned int)rngCurrent ) {
 	  UserCommand(set_rng_cmds[i]);
-  } else if (i == RNG_FILE && i != rngCurrent ) {
+  } else if (i == RNG_FILE && i != (unsigned int)rngCurrent ) {
       filename = GTKFileSelect (_("Select file with dice"), NULL, NULL, NULL,
 			      GTK_FILE_CHOOSER_ACTION_OPEN);
       if (filename)
